支持自动重新编译和加载插件

+++ ZKWeb/Manager/PluginManager.cs
@@ -35,6 +35,7 @@ namespace ZKWeb.Manager {
 		public PluginManager() {
 			// 获取网站配置中的插件列表
 			var configManager = Application.Current.Ioc.Resolve<ConfigManager>();
+			var logManager = Application.Current.Ioc.Resolve<LogManager>();
 			// 载入所有插件信息
 			foreach (var pluginName in configManager.WebsiteConfig.Plugins) {
 				var dir = PathUtils.SecureCombine(PathConfig.PluginsRootDirectory, pluginName);
@@ -72,6 +73,7 @@ namespace ZKWeb.Manager {
 				}
 			}
 			// 查找插件的引用程序集目录
+			// 这里不查找插件的程序集目录避免错误的在重新编译前载入了插件的程序集
 			foreach (var plugin in Plugins) {
 				var path = Path.Combine(plugin.ReferencesDirectory(), $"{requireName.Name}.dll");
 				if (File.Exists(path)) {
@@ -92,13 +94,21 @@ namespace ZKWeb.Manager {
 		/// </summary>
 		/// <param name="info">插件信息</param>
 		public static void Compile(this PluginInfo info) {
-			// 检查是否需要重新编译
-			// 没有源文件时表示可能是只有资源文件或不开源的插件，这时不需要重新编译
-			// TODO 比较源文件的修改时间
+			// 获取插件的源代码文件列表和各个路径
 			var sourceFiles = info.SourceFiles();
 			var assemblyPath = info.AssemblyPath();
 			var assemblyPdbPath = info.AssemblyPdbPath();
-			if (sourceFiles.Length > 0) {
+			var compileInfoPath = info.CompileInfoPath();
+			// 检查是否需要重新编译
+			// 会通过对比所有源文件的修改时间是否一致来检查
+			// 没有源文件时表示只有资源文件或不开源，不需要重新编译
+			var existCompileInfo = "";
+			if (File.Exists(compileInfoPath)) {
+				existCompileInfo = File.ReadAllText(compileInfoPath);
+			}
+			var compileInfo = string.Join("\r\n",
+				sourceFiles.OrderBy(s => s).Select(s => $"{s} {File.GetLastWriteTime(s)}"));
+			if (sourceFiles.Length > 0 && compileInfo != existCompileInfo) {
 				// 重新编译前把原来的文件重命名为old文件
 				if (File.Exists(assemblyPath)) {
 					File.Move(assemblyPath, $"{assemblyPath}.{DateTime.UtcNow.Ticks}.old");
@@ -109,11 +119,23 @@ namespace ZKWeb.Manager {
 				// 编译，debug = true可以生成pdb文件以支持断点
 				Directory.CreateDirectory(info.BinDirectory());
 				CSScript.CompileFiles(sourceFiles, assemblyPath, true, info.References);
+				// 保存编译信息
+				File.WriteAllText(compileInfoPath, compileInfo);
 				// 删除old文件
 				// 有可能因为文件占用而删除不成功，忽略删除失败时的例外
 				Directory.EnumerateFiles(info.BinDirectory(), "*.old").ForEach(
 					path => { try { File.Delete(path); } catch { } });
 			}
+		}
+
+		/// <summary>
+		/// 获取插件目录的名称
+		/// 例 Common.Base
+		/// </summary>
+		/// <param name="info">插件信息</param>
+		/// <returns></returns>
+		public static string DirectoryName(this PluginInfo info) {
+			return Path.GetFileName(info.Directory);
 		}
 
 		/// <summary>
@@ -162,7 +184,7 @@ namespace ZKWeb.Manager {
 		/// <param name="info">插件信息</param>
 		/// <returns></returns>
 		public static string AssemblyPath(this PluginInfo info) {
-			return Path.Combine(info.BinDirectory(), $"{Path.GetFileName(info.Directory)}.dll");
+			return Path.Combine(info.BinDirectory(), $"{info.DirectoryName()}.dll");
 		}
 
 		/// <summary>
@@ -171,7 +193,18 @@ namespace ZKWeb.Manager {
 		/// <param name="info">插件信息</param>
 		/// <returns></returns>
 		public static string AssemblyPdbPath(this PluginInfo info) {
-			return Path.Combine(info.BinDirectory(), $"{Path.GetFileName(info.Directory)}.pdb");
+			return Path.Combine(info.BinDirectory(), $"{info.DirectoryName()}.pdb");
+		}
+
+		/// <summary>
+		/// 获取保存编译信息的文件路径
+		/// 这个文件中有源代码和修改时间的列表
+		/// 用于检测是否需要重新编译
+		/// </summary>
+		/// <param name="info">插件信息</param>
+		/// <returns></returns>
+		public static string CompileInfoPath(this PluginInfo info) {
+			return Path.Combine(info.BinDirectory(), "CompileInfo.txt");
 		}
 	}
 }
